Validaciones hechas, forms de usuario listos, les falta validar

+++ SistemaMLP/Forms/UserForms/FrmUsers.cs
@@ -1,5 +1,6 @@
 ﻿using MLPlib.Class;
 using System;
+using System.Linq;
 using System.Windows.Forms;
 
 namespace SistemaMLP.Forms.UserForms
@@ -7,6 +8,7 @@ namespace SistemaMLP.Forms.UserForms
     public partial class FrmUsers : Form
     {
         public User user = new User();
+        public bool DeletedMode = false;
         public FrmUsers()
         {
             InitializeComponent();
@@ -14,7 +16,28 @@ namespace SistemaMLP.Forms.UserForms
 
         private void TxtSearch_TextChanged(object sender, EventArgs e)
         {
-
+            if (!DeletedMode)
+            {
+                if (!string.IsNullOrEmpty(TxtSearch.Text.Trim()) && TxtSearch.Text.Count() >= 1)
+                {
+                    FillDGV(TxtSearch.Text.Trim());
+                }
+                else
+                {
+                    FillDGV();
+                }
+            }
+            else
+            {
+                if (!string.IsNullOrEmpty(TxtSearch.Text.Trim()) && TxtSearch.Text.Count() >= 1)
+                {
+                    FillDeletedDGV(TxtSearch.Text.Trim());
+                }
+                else
+                {
+                    FillDeletedDGV();
+                }
+            }
         }
 
         private void FrmUsers_Load(object sender, EventArgs e)
@@ -38,12 +61,29 @@ namespace SistemaMLP.Forms.UserForms
 
         }
 
+        private void FillDeletedDGV(string filter = "")
+        {
+            DGVUsers.DataSource = user.GetDeletedUsers();
+
+            DGVUsers.Columns["IDUser"].Visible = false;
+            DGVUsers.Columns["Active"].Visible = false;
+
+            DGVUsers.Columns["Username"].HeaderText = "Nombre de usuario";
+            DGVUsers.Columns["RegDate"].HeaderText = "Fecha de registro";
+
+            DGVUsers.Columns["Username"].AutoSizeMode = DataGridViewAutoSizeColumnMode.Fill;
+            DGVUsers.Columns["RegDate"].AutoSizeMode = DataGridViewAutoSizeColumnMode.Fill;
+
+        }
+
         private void CleanForm()
         {
             TxtSearch.Text = "";
             DGVUsers.ClearSelection();
             BtnEdit.Enabled = false;
             BtnDelete.Enabled = false;
+            CbDeleted.Checked = false;
+            BtnLayout();
         }
 
         private void TxtSearch_MouseHover(object sender, EventArgs e)
@@ -100,9 +140,36 @@ namespace SistemaMLP.Forms.UserForms
 
         private void DGVUsers_CellClick(object sender, DataGridViewCellEventArgs e)
         {
-            if (DGVUsers.SelectedRows.Count > 0)
+            if (!DeletedMode)
             {
+                if (DGVUsers.SelectedRows.Count > 0)
+                {
+                    BtnAdd.Enabled = true;
+                    BtnEdit.Enabled = true;
+                    BtnDelete.Enabled = true;
+                }
+            }
+            else
+            {
+                if (DGVUsers.SelectedRows.Count > 0)
+                {
+                    BtnLayout();
+                }
+            }
+        }
+
+        private void BtnLayout()
+        {
+            if (!DeletedMode)
+            {
+                BtnAdd.Enabled = true;
                 BtnEdit.Enabled = true;
+                BtnDelete.Enabled = false;
+            }
+            else
+            {
+                BtnAdd.Enabled = false;
+                BtnEdit.Enabled = false;
                 BtnDelete.Enabled = true;
             }
         }
@@ -116,27 +183,85 @@ namespace SistemaMLP.Forms.UserForms
         {
             if (DGVUsers.SelectedRows.Count > 0)
             {
-                DialogResult dialogResult = MessageBox.Show("Está seguro que desea eliminar ese usuario?", "Eliminar usuario", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-                if (dialogResult == DialogResult.Yes)
+                if (!DeletedMode)
                 {
-                    DataGridViewRow dataRow = DGVUsers.SelectedRows[0];
-
-                    user = new User
+                    DialogResult dialogResult = MessageBox.Show("Está seguro que desea eliminar ese usuario?", "Eliminar usuario", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                    if (dialogResult == DialogResult.Yes)
                     {
-                        IDUser = Convert.ToInt32(dataRow.Cells["IDUser"].Value),
+                        DataGridViewRow dataRow = DGVUsers.SelectedRows[0];
 
-                    };
-                    if (user.DeleteUser() == 1)
-                    {
-                        MessageBox.Show("Se ha eliminado el usuario", "Usuario eliminado", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                        CleanForm();
+                        user = new User
+                        {
+                            IDUser = Convert.ToInt32(dataRow.Cells["IDUser"].Value),
+                            Username = Convert.ToString(dataRow.Cells["Username"].Value),
+
+                        };
+                        if(user.IDUser != Utilities.Utilities.user.IDUser)
+                        {
+                            if (user.DeleteUser() == 1)
+                            {
+                                MessageBox.Show("Se ha eliminado el usuario", "Usuario eliminado", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                                Utilities.Utilities.CreateLog("ha eliminado el usuario: " + user.Username);
+                                CleanForm();
+                                FillDGV();
+                            }
+                            else
+                            {
+                                MessageBox.Show("Ha ocurrido un error al intentar eliminar el usuario", "Error de eliminado", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                            }
+                        }
+                        else
+                        {
+                            MessageBox.Show("No se permite eliminar el usuario en sesión", "Error de eliminado", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                            return;
+                        }
                     }
-                    else
+                }
+                else
+                {
+                    DialogResult dialogResult = MessageBox.Show("Está seguro que desea restaurar ese usuario?", "Restaurar usuario", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                    if (dialogResult == DialogResult.Yes)
                     {
-                        MessageBox.Show("Ha ocurrido un error al intentar eliminar el usuario", "Error de eliminado", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                        DataGridViewRow dataRow = DGVUsers.SelectedRows[0];
+
+                        user = new User
+                        {
+                            IDUser = Convert.ToInt32(dataRow.Cells["IDUser"].Value),
+                            Username = Convert.ToString(dataRow.Cells["Username"].Value),
+
+                        };
+                        if (user.RestoreUser() == 1)
+                        {
+                            MessageBox.Show("Se ha restaurado el usuario", "Usuario restaurado", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                            Utilities.Utilities.CreateLog("ha restaurado el usuario: " + user.Username);
+                            CleanForm();
+                            FillDGV();
+                        }
+                        else
+                        {
+                            MessageBox.Show("Ha ocurrido un error al intentar eliminar el usuario", "Error de eliminado", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                        }
                     }
                 }
 
+            }
+        }
+
+        private void CbDeleted_CheckedChanged(object sender, EventArgs e)
+        {
+            if(CbDeleted.Checked == true)
+            {
+                DeletedMode = true;
+                BtnDelete.Text = "Restaurar usuario";
+                FillDeletedDGV();
+                BtnLayout();
+            }
+            else
+            {
+                DeletedMode = false;
+                BtnDelete.Text = "Eliminar usuario";
+                FillDGV();
+                BtnLayout();
             }
         }
     }
